Code Refractoring + new files
